requirements.txt
python-telegram-bot==20.7
firebase-admin
matplotlib
python-dotenv